+++ web/src/pages/Archived.tsx
@@ -1,21 +1,20 @@
 import { Button, Tooltip } from "@mui/joy";
 import { ClientError } from "nice-grpc-web";
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useState } from "react";
 import toast from "react-hot-toast";
-import { showCommonDialog } from "@/components/Dialog/CommonDialog";
 import Empty from "@/components/Empty";
 import Icon from "@/components/Icon";
 import MemoContent from "@/components/MemoContent";
-import MemoFilter from "@/components/MemoFilter";
+import MemoFilters from "@/components/MemoFilters";
+import MemoResourceListView from "@/components/MemoResourceListView";
 import MobileHeader from "@/components/MobileHeader";
 import SearchBar from "@/components/SearchBar";
 import { DEFAULT_LIST_MEMOS_PAGE_SIZE } from "@/helpers/consts";
 import { getTimeStampByDate } from "@/helpers/datetime";
 import useCurrentUser from "@/hooks/useCurrentUser";
-import useFilterWithUrlParams from "@/hooks/useFilterWithUrlParams";
-import { useMemoList, useMemoStore } from "@/store/v1";
-import { RowStatus } from "@/types/proto/api/v2/common";
-import { Memo } from "@/types/proto/api/v2/memo_service";
+import { useMemoFilterStore, useMemoList, useMemoStore } from "@/store/v1";
+import { RowStatus } from "@/types/proto/api/v1/common";
+import { Memo } from "@/types/proto/api/v1/memo_service";
 import { useTranslate } from "@/utils/i18n";
 
 const Archived = () => {
@@ -23,51 +22,50 @@ const Archived = () => {
   const user = useCurrentUser();
   const memoStore = useMemoStore();
   const memoList = useMemoList();
+  const memoFilterStore = useMemoFilterStore();
   const [isRequesting, setIsRequesting] = useState(true);
-  const nextPageTokenRef = useRef<string | undefined>(undefined);
-  const { tag: tagQuery, text: textQuery } = useFilterWithUrlParams();
+  const [nextPageToken, setNextPageToken] = useState<string>("");
   const sortedMemos = memoList.value
     .filter((memo) => memo.rowStatus === RowStatus.ARCHIVED)
     .sort((a, b) => getTimeStampByDate(b.displayTime) - getTimeStampByDate(a.displayTime));
 
   useEffect(() => {
-    nextPageTokenRef.current = undefined;
     memoList.reset();
-    fetchMemos();
-  }, [tagQuery, textQuery]);
+    fetchMemos("");
+  }, [memoFilterStore.filters]);
 
-  const fetchMemos = async () => {
+  const fetchMemos = async (nextPageToken: string) => {
+    setIsRequesting(true);
     const filters = [`creator == "${user.name}"`, `row_status == "ARCHIVED"`];
     const contentSearch: string[] = [];
-    if (tagQuery) {
-      contentSearch.push(JSON.stringify(`#${tagQuery}`));
-    }
-    if (textQuery) {
-      contentSearch.push(JSON.stringify(textQuery));
+    const tagSearch: string[] = [];
+    for (const filter of memoFilterStore.filters) {
+      if (filter.factor === "contentSearch") {
+        contentSearch.push(`"${filter.value}"`);
+      } else if (filter.factor === "tagSearch") {
+        tagSearch.push(`"${filter.value}"`);
+      }
     }
     if (contentSearch.length > 0) {
       filters.push(`content_search == [${contentSearch.join(", ")}]`);
     }
-    setIsRequesting(true);
-    const data = await memoStore.fetchMemos({
+    if (tagSearch.length > 0) {
+      filters.push(`tag_search == [${tagSearch.join(", ")}]`);
+    }
+    const response = await memoStore.fetchMemos({
       pageSize: DEFAULT_LIST_MEMOS_PAGE_SIZE,
       filter: filters.join(" && "),
-      pageToken: nextPageTokenRef.current,
+      pageToken: nextPageToken,
     });
     setIsRequesting(false);
-    nextPageTokenRef.current = data.nextPageToken;
+    setNextPageToken(response.nextPageToken);
   };
 
   const handleDeleteMemoClick = async (memo: Memo) => {
-    showCommonDialog({
-      title: t("memo.delete-memo"),
-      content: t("memo.delete-confirm"),
-      style: "danger",
-      dialogName: "delete-memo-dialog",
-      onConfirm: async () => {
-        await memoStore.deleteMemo(memo.name);
-      },
-    });
+    const confirmed = window.confirm(t("memo.delete-confirm"));
+    if (confirmed) {
+      await memoStore.deleteMemo(memo.name);
+    }
   };
 
   const handleRestoreMemoClick = async (memo: Memo) => {
@@ -91,12 +89,16 @@ const Archived = () => {
       <MobileHeader />
       <div className="w-full px-4 sm:px-6">
         <div className="w-full flex flex-col justify-start items-start">
-          <div className="w-full flex flex-row justify-end items-center mb-2">
-            <div className="w-40">
+          <div className="w-full flex flex-row justify-between items-center mb-2">
+            <div className="flex flex-row justify-start items-center gap-1">
+              <Icon.Archive className="w-5 h-auto opacity-70 shrink-0" />
+              <span>{t("common.archived")}</span>
+            </div>
+            <div className="w-44">
               <SearchBar />
             </div>
           </div>
-          <MemoFilter className="px-2 pb-2" />
+          <MemoFilters />
           {sortedMemos.map((memo) => (
             <div
               key={memo.name}
@@ -121,7 +123,8 @@ const Archived = () => {
                   </Tooltip>
                 </div>
               </div>
-              <MemoContent key={`${memo.name}-${memo.displayTime}`} memoName={memo.name} content={memo.content} readonly={true} />
+              <MemoContent key={`${memo.name}-${memo.displayTime}`} memoName={memo.name} nodes={memo.nodes} readonly={true} />
+              <MemoResourceListView resources={memo.resources} />
             </div>
           ))}
           {isRequesting ? (
@@ -129,7 +132,7 @@ const Archived = () => {
               <Icon.Loader className="w-4 h-auto animate-spin mr-1" />
               <p className="text-sm italic">{t("memo.fetching-data")}</p>
             </div>
-          ) : !nextPageTokenRef.current ? (
+          ) : !nextPageToken ? (
             sortedMemos.length === 0 && (
               <div className="w-full mt-16 mb-8 flex flex-col justify-center items-center italic">
                 <Empty />
@@ -138,7 +141,7 @@ const Archived = () => {
             )
           ) : (
             <div className="w-full flex flex-row justify-center items-center my-4">
-              <Button variant="plain" endDecorator={<Icon.ArrowDown className="w-5 h-auto" />} onClick={fetchMemos}>
+              <Button variant="plain" endDecorator={<Icon.ArrowDown className="w-5 h-auto" />} onClick={() => fetchMemos(nextPageToken)}>
                 {t("memo.fetch-more")}
               </Button>
             </div>
