+++ web/src/pages/Resources.tsx
@@ -1,8 +1,6 @@
 import { Divider, IconButton, Input, Tooltip } from "@mui/joy";
 import { includes } from "lodash-es";
 import { useEffect, useState } from "react";
-import { Link } from "react-router-dom";
-import { showCommonDialog } from "@/components/Dialog/CommonDialog";
 import Empty from "@/components/Empty";
 import Icon from "@/components/Icon";
 import MobileHeader from "@/components/MobileHeader";
@@ -10,29 +8,28 @@ import ResourceIcon from "@/components/ResourceIcon";
 import { resourceServiceClient } from "@/grpcweb";
 import useLoading from "@/hooks/useLoading";
 import i18n from "@/i18n";
-import { extractMemoIdFromName, useMemoStore } from "@/store/v1";
-import { Resource } from "@/types/proto/api/v2/resource_service";
+import { useMemoStore } from "@/store/v1";
+import { Resource } from "@/types/proto/api/v1/resource_service";
 import { useTranslate } from "@/utils/i18n";
 
 function groupResourcesByDate(resources: Resource[]) {
-  const tmp_resources: Resource[] = resources.slice();
-  tmp_resources.sort((a: Resource, b: Resource) => {
-    const a_date = new Date(a.createTime as any);
-    const b_date = new Date(b.createTime as any);
-    return b_date.getTime() - a_date.getTime();
-  });
-
   const grouped = new Map<number, Resource[]>();
-  tmp_resources.forEach((item) => {
-    const date = new Date(item.createTime as any);
-    const year = date.getFullYear();
-    const month = date.getMonth() + 1;
-    const timestamp = Date.UTC(year, month - 1, 1);
-    if (!grouped.has(timestamp)) {
-      grouped.set(timestamp, []);
-    }
-    grouped.get(timestamp)?.push(item);
-  });
+  resources
+    .sort((a: Resource, b: Resource) => {
+      const a_date = new Date(a.createTime as any);
+      const b_date = new Date(b.createTime as any);
+      return b_date.getTime() - a_date.getTime();
+    })
+    .forEach((item) => {
+      const date = new Date(item.createTime as any);
+      const year = date.getFullYear();
+      const month = date.getMonth() + 1;
+      const timestamp = Date.UTC(year, month - 1, 1);
+      if (!grouped.has(timestamp)) {
+        grouped.set(timestamp, []);
+      }
+      grouped.get(timestamp)?.push(item);
+    });
   return grouped;
 }
 
@@ -60,19 +57,14 @@ const Resources = () => {
     });
   }, []);
 
-  const handleDeleteUnusedResources = () => {
-    showCommonDialog({
-      title: "Delete all unused resources",
-      content: "Are you sure to delete all unused resources? This action cannot be undone.",
-      style: "warning",
-      dialogName: "delete-unused-resources-dialog",
-      onConfirm: async () => {
-        for (const resource of unusedResources) {
-          await resourceServiceClient.deleteResource({ name: resource.name });
-        }
-        setResources(resources.filter((resource) => resource.memo));
-      },
-    });
+  const handleDeleteUnusedResources = async () => {
+    const confirmed = window.confirm("Are you sure to delete all unused resources? This action cannot be undone.");
+    if (confirmed) {
+      for (const resource of unusedResources) {
+        await resourceServiceClient.deleteResource({ name: resource.name });
+      }
+      setResources(resources.filter((resource) => resource.memo));
+    }
   };
 
   return (
@@ -119,7 +111,6 @@ const Resources = () => {
                           </div>
                           <div className="w-full max-w-[calc(100%-4rem)] sm:max-w-[calc(100%-6rem)] flex flex-row justify-start items-start gap-4 flex-wrap">
                             {resources.map((resource) => {
-                              const relatedMemo = resource.memo ? memoStore.getMemoByName(resource.memo) : null;
                               return (
                                 <div key={resource.name} className="w-24 sm:w-32 h-auto flex flex-col justify-start items-start">
                                   <div className="w-24 h-24 flex justify-center items-center sm:w-32 sm:h-32 border dark:border-zinc-900 overflow-clip rounded-xl cursor-pointer hover:shadow hover:opacity-80">
@@ -127,14 +118,6 @@ const Resources = () => {
                                   </div>
                                   <div className="w-full max-w-full flex flex-row justify-between items-center mt-1 px-1">
                                     <p className="text-xs shrink text-gray-400 truncate">{resource.filename}</p>
-                                    {relatedMemo && (
-                                      <Link
-                                        className="shrink-0 text-xs ml-1 text-gray-400 hover:underline hover:text-blue-600"
-                                        to={`/m/${relatedMemo.uid}`}
-                                      >
-                                        #{extractMemoIdFromName(relatedMemo.name)}
-                                      </Link>
-                                    )}
                                   </div>
                                 </div>
                               );
